Iteration 6: TypeScript Refinement
